Move MetaMapBuilder into its own file
